Gsictl

See merge request docs/templates/makefiles!3

+++ git.mk
@@ -1,19 +1,28 @@
-git_commit:
+commit:
 	git commit -a
 
-git_tag:
+tag:
 	git tag ${VERSION}
 
-git_push:
-	git push --tags origin master
+git-push::
+	git push --tags -u origin HEAD
 
-push-github: ## Push the code to github. You need to set up HUB_USER and HUB_PASSWORD
+git-pull:
+	git pull --all
+
+push-github: ## Push the code to github. You need to set up GITHUB_DEPLOY_KEY
+ifeq ($(GITHUB_DEPLOY_KEY),)
+else
 	$(eval KEY_FILE := $(shell mktemp))
-	@echo "$$GITHUB_DEPLOY_KEY" > $(KEY_FILE)
+	@echo "$(GITHUB_DEPLOY_KEY)" > $(KEY_FILE)
 	@git remote rm github-deploy || true
 	git remote add github-deploy $(GITHUB_REPO)
 	@GIT_SSH_COMMAND="ssh -i $(KEY_FILE)" git fetch github-deploy $(CI_COMMIT_REF_NAME) || true
 	@GIT_SSH_COMMAND="ssh -i $(KEY_FILE)" git push github-deploy $(CI_COMMIT_REF_NAME)
 	rm $(KEY_FILE)
+endif
 
-.PHONY:: git_commit git_tag git_push push-github
+push:: git-push
+pull:: git-pull
+
+.PHONY:: commit tag push git-push git-pull push-github
